% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_functions.R
\name{creat_cyjFI}
\alias{creat_cyjFI}
\title{Creat a cytoscape.js shiny based on the}
\usage{
creat_cyjFI(PPI_res, Annotation_enrichment_res, DEP_res)
}
\arguments{
\item{PPI_res}{The result from test_ppi or test_reactome}

\item{Annotation_enrichment_res}{A enrichResult object (or its data.frame result) that stores enrichment analysis result}

\item{DEP_res}{The dep result from DEP2/DEP}
}
\value{
A PPI.enrich.R6.Shiny R6 class object, containing the analysis results and shiny application
}
\description{
Creat a cytoscape.js shiny based on the
}
