% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_complex.R
\name{enrichComplex}
\alias{enrichComplex}
\title{Enrich REACTOME complex}
\usage{
enrichComplex(proteinlist)
}
\arguments{
\item{proteinlist}{Vector, human uniprot protein ID}
}
\description{
Base on clusterProfiler::enricher, enrich REACTOME complex on giving protein list.
}
