% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6class.R
\name{PPI.enrich.R6.Shiny}
\alias{PPI.enrich.R6.Shiny}
\title{R6 Class store cyjFI}
\description{
This R6 object store a shiny app to visual display protein interaction and enriched protein annotation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PPI.enrich.R6.Shiny-return_cyj}{\code{PPI.enrich.R6.Shiny$return_cyj()}}
\item \href{#method-PPI.enrich.R6.Shiny-return_argument}{\code{PPI.enrich.R6.Shiny$return_argument()}}
\item \href{#method-PPI.enrich.R6.Shiny-return_session}{\code{PPI.enrich.R6.Shiny$return_session()}}
\item \href{#method-PPI.enrich.R6.Shiny-return_network_json}{\code{PPI.enrich.R6.Shiny$return_network_json()}}
\item \href{#method-PPI.enrich.R6.Shiny-new}{\code{PPI.enrich.R6.Shiny$new()}}
\item \href{#method-PPI.enrich.R6.Shiny-run_app}{\code{PPI.enrich.R6.Shiny$run_app()}}
\item \href{#method-PPI.enrich.R6.Shiny-ui}{\code{PPI.enrich.R6.Shiny$ui()}}
\item \href{#method-PPI.enrich.R6.Shiny-server}{\code{PPI.enrich.R6.Shiny$server()}}
\item \href{#method-PPI.enrich.R6.Shiny-clone}{\code{PPI.enrich.R6.Shiny$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-return_cyj"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-return_cyj}{}}}
\subsection{Method \code{return_cyj()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$return_cyj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-return_argument"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-return_argument}{}}}
\subsection{Method \code{return_argument()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$return_argument()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-return_session"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-return_session}{}}}
\subsection{Method \code{return_session()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$return_session()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-return_network_json"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-return_network_json}{}}}
\subsection{Method \code{return_network_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$return_network_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-new"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$new(PPI_res = NA, ORA_res = NA, DEP_res = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-run_app"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-run_app}{}}}
\subsection{Method \code{run_app()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$run_app(port = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-ui"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-ui}{}}}
\subsection{Method \code{ui()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$ui()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-server"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$server(
  input,
  output,
  session,
  PPIres = self$PPIres,
  ORAres = self$ORAres,
  DEPres = self$DEPres
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PPI.enrich.R6.Shiny-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PPI.enrich.R6.Shiny-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PPI.enrich.R6.Shiny$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
