% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPI_functions.R
\name{test_ReactomeFI}
\alias{test_ReactomeFI}
\title{Title}
\usage{
test_ReactomeFI(
  x,
  contrasts = NULL,
  species = "Human",
  reactome_FI_data = NULL,
  score_cutoff = 0.5
)
}
\arguments{
\item{x}{A SummarizedExperiment/DEGdata output from \code{\link[DEP2]{add_rejections}} or
a charachter vector containing candidate identifiers(SYMBOL, EntrezID, UniprotID or ENSEMBL).}

\item{contrasts}{Character, specifies the contrasts to get significant candidates.}

\item{species}{The species name.}

\item{score_cutoff}{A numeric less than 1, required lowest interaction scores. A score lowwer than 400 means the interaction is unconfident}
}
\description{
Title
}
\examples{
data()
}
